requirements.txt
fastapi==0.95.0
uvicorn==0.23.1
opencv-python-headless==4.7.0.72
numpy==1.25.1
Pillow==10.0.0
ultralytics==8.2.86
python-multipart
line-bot-sdk